Google Docs Sheet order incompatibility
Fixes #13.

Google Docs does not emit worksheet information, and furthermore does not follow
the sheet order as prescribed in the `[Content-Types].xml` metadata.  This
workaround forces it to treat the sheet list in Workbook as the map to the files
`xl/worksheets/sheet##.xml` even though the rIds are different.