readAsArrayBuffer compatibility and fmt on user hand
First, I changed `firstbyte()` to verify if 'f' is an ArrayBuffer (when using `readAsArrayBuffer` instead of `readAsBinaryString`) and if it's the case it will change it to a string.

Second, I changed `safe_format()` to include `fmt` and `fmtid` into the object returned to the user. That way, each person will be able to handle the format as required.