Support of cellDates
Currently the `cellDates` option doesn't work and the Date cells are not correctly recognized.

I propose a fix that returns `.d` that contains the JavaScript Date.

I also propose to provide `fmt` and `fmtid` to the object so the user will be able to deal with any customized formats. It will give more flexibility.