fix: #1717
I saw the issue was raised and a SheetJS Dev said the fix was simple and involved dealing with defined and undefined variables.
They wrote a line using a ternary conditional to assign the variable textp to the value of cell.v if it's defined, or an empty string if it is not defined. 

SheetJS Dev Code: textp = cell.v == null ? "" : cell.v;

Suggesting using the line: textp = cell.v || "";

This does the same thing and assigns textp to cell.v if it is defined or an empty string if it is not. 

(Note any feedback would be greatly appreciated, this is my first try at contributing to open source).